# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "医·基层AI辅助诊断系统 API",
    "version": "1.45.0",
    "description": "Cloudflare Pages Functions（线上权威实现 functions/api/[[route]].js）的机器可读契约；FastAPI 后端为镜像实现，双端一致性由 contract_parity 测试保障。所有响应统一壳 {code:0,data}（成功）或 {code:<http-status>,message}（失败），并携带 x-request-id 响应头用于故障归因。全程 AI 辅助参考 · 医生终审，输出不构成独立诊断。",
    "x-compliance": "演示仅脱敏合成病例；evidence_ids 仅限知识库白名单 kb-001~055；红旗规则层独立于 LLM 不可被模型覆盖"
  },
  "servers": [
    {
      "url": "https://doctor-ai-dx.pages.dev",
      "description": "生产（Pages Functions）"
    },
    {
      "url": "http://127.0.0.1:8000",
      "description": "本地 FastAPI 镜像（run.py）"
    }
  ],
  "paths": {
...
      "/api/chat": {
        "post": {
          "tags": [
            "chat"
          ],
          "summary": "对话入口：红旗闸门 → 意图识别 → FAQ 检索 → 转人工判定 → 脱敏落库",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "text"
                  ],
                  "properties": {
                    "text": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 2000
                    },
                    "history": {
                      "type": "array",
                      "maxItems": 64,
                      "items": {
                        "type": "object",
                        "properties": {
                          "role": {
                            "type": "string"
                          },
                          "content": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "conversation_id": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "对话结果；mode=deterministic（判定全在确定性代码，红旗不经模型）",
              "headers": {
                "x-request-id": {
                  "$ref": "#/components/headers/RequestId"
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Envelope"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "data": {
                            "$ref": "#/components/schemas/ChatResult"
                          }
                        }
                      }
                    ]
                  }
                }
              }
            },
            "400": {
              "$ref": "#/components/responses/BadJson"
            },
            "404": {
              "$ref": "#/components/responses/NotFound"
            },
            "413": {
              "$ref": "#/components/responses/PayloadTooLarge"
            },
            "422": {
              "$ref": "#/components/responses/BadShape"
            },
            "500": {
              "$ref": "#/components/responses/InternalError"
            }
          }
        }
      },
      "/api/chat/{conversationId}": {
        "get": {
          "tags": [
            "chat"
          ],
          "summary": "取单个会话详情（脱敏视图）",
          "parameters": [
            {
              "name": "conversationId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "成功",
              "headers": {
                "x-request-id": {
                  "$ref": "#/components/headers/RequestId"
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Envelope"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "data": {
                            "$ref": "#/components/schemas/ConversationDetail"
                          }
                        }
                      }
                    ]
                  }
                }
              }
            },
            "404": {
              "$ref": "#/components/responses/NotFound"
            },
            "500": {
              "$ref": "#/components/responses/InternalError"
            }
          }
        }
      },
      "/api/chat/{conversationId}/feedback": {
        "post": {
          "tags": [
            "chat"
          ],
          "summary": "会话满意度打分（1-5，越界拒绝不静默截断）",
          "parameters": [
            {
              "name": "conversationId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "score"
                  ],
                  "properties": {
                    "score": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 5
                    },
                    "tag": {
                      "type": "string",
                      "maxLength": 40
                    },
                    "comment": {
                      "type": "string",
                      "maxLength": 200
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "已记录",
              "headers": {
                "x-request-id": {
                  "$ref": "#/components/headers/RequestId"
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Envelope"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "data": {
                            "type": "object"
                          }
                        }
                      }
                    ]
                  }
                }
              }
            },
            "400": {
              "$ref": "#/components/responses/BadJson"
            },
            "404": {
              "$ref": "#/components/responses/NotFound"
            },
            "413": {
              "$ref": "#/components/responses/PayloadTooLarge"
            },
            "422": {
              "$ref": "#/components/responses/BadShape"
            },
            "500": {
              "$ref": "#/components/responses/InternalError"
            }
          }
        }
      },
      "/api/admin/conversations": {
        "get": {
          "tags": [
            "admin"
          ],
          "summary": "管理后台：会话列表（需 x-admin-token）",
          "parameters": [
            {
              "name": "limit",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "minimum": 1,
                "maximum": 100
              }
            },
            {
              "name": "offset",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "minimum": 0
              }
            }
          ],
          "responses": {
            "200": {
              "description": "成功",
              "headers": {
                "x-request-id": {
                  "$ref": "#/components/headers/RequestId"
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Envelope"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "data": {
                            "type": "object"
                          }
                        }
                      }
                    ]
                  }
                }
              }
            },
            "401": {
              "description": "未提供管理令牌"
            },
            "403": {
              "description": "管理令牌不正确"
            },
            "500": {
              "$ref": "#/components/responses/InternalError"
            },
            "503": {
              "description": "服务器未配置 ADMIN_TOKEN"
            }
          }
        }
      },
...
      "/api/admin/handoffs": {
        "get": {
          "tags": [
            "admin"
          ],
          "summary": "管理后台：转人工队列（需 x-admin-token）",
          "parameters": [
            {
              "name": "limit",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "minimum": 1,
                "maximum": 100
              }
            }
          ],
          "responses": {
            "200": {
              "description": "成功",
              "headers": {
                "x-request-id": {
                  "$ref": "#/components/headers/RequestId"
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Envelope"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "data": {
                            "type": "object"
                          }
                        }
                      }
                    ]
                  }
                }
              }
            },
            "401": {
              "description": "未提供管理令牌"
            },
            "403": {
              "description": "管理令牌不正确"
            },
            "500": {
              "$ref": "#/components/responses/InternalError"
            },
            "503": {
              "description": "服务器未配置 ADMIN_TOKEN"
            }
          }
        }
      },
      "/api/admin/stats": {
        "get": {
          "tags": [
            "admin"
          ],
          "summary": "管理后台：满意度与意图统计（需 x-admin-token）",
          "responses": {
            "200": {
              "description": "成功",
              "headers": {
                "x-request-id": {
                  "$ref": "#/components/headers/RequestId"
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Envelope"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "data": {
                            "type": "object"
                          }
                        }
                      }
                    ]
                  }
                }
              }
            },
            "401": {
              "description": "未提供管理令牌"
            },
            "403": {
              "description": "管理令牌不正确"
            },
            "500": {
              "$ref": "#/components/responses/InternalError"
            },
            "503": {
              "description": "服务器未配置 ADMIN_TOKEN"
            }
          }
        }
      },
...
    "/api/intake/ask": {
      "post": {
        "tags": [
          "clinical"
        ],
        "summary": "智能问诊推进（下一追问或置 done；含 state 临床状态抽取）",
        "requestBody": {
          "$ref": "#/components/requestBodies/IntakeAsk"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonEnvelope"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "400": {
            "$ref": "#/components/responses/BadJson"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "422": {
            "$ref": "#/components/responses/BadShape"
          }
        }
      }
    },
    "/api/dx/{caseId}": {
      "post": {
        "tags": [
          "clinical"
        ],
        "summary": "辅助诊断（状态抽取→BM25 检索→LLM 生成→确定性校验→红旗兜底→失败降级）",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "c1"
            }
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/HistoryDx"
        },
        "responses": {
          "200": {
            "description": "诊断结果；mode=live|rule-fallback；红旗不可被模型覆盖",
            "headers": {
              "x-request-id": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Diagnosis"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "400": {
            "$ref": "#/components/responses/BadJson"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "422": {
            "$ref": "#/components/responses/BadShape"
          }
        }
      }
    },
    "/api/workup/{caseId}": {
      "post": {
        "tags": [
          "clinical"
        ],
        "summary": "检查建议（essential/suggested/optional 三组；可复用已生成诊断省一次串行 LLM）",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/HistoryDx"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonEnvelope"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "400": {
            "$ref": "#/components/responses/BadJson"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "422": {
            "$ref": "#/components/responses/BadShape"
          }
        }
      }
    },
    "/api/report/{caseId}": {
      "post": {
        "tags": [
          "clinical"
        ],
        "summary": "SOAP 病历报告（含免责声明与医生终审定位）",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/HistoryDx"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonEnvelope"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "400": {
            "$ref": "#/components/responses/BadJson"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "422": {
            "$ref": "#/components/responses/BadShape"
          }
        }
      }
    },
...
  },
  "components": {
    "headers": {
      "RequestId": {
        "description": "请求归因编号（故障反馈凭据；不回显堆栈/内部路径）",
        "schema": {
          "type": "string"
        }
      }
    },
    "requestBodies": {
      "IntakeAsk": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "case_id"
              ],
              "properties": {
                "case_id": {
                  "type": "string"
                },
                "history": {
                  "$ref": "#/components/schemas/History"
                }
              }
            }
          }
        }
      },
      "HistoryDx": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "history": {
                  "$ref": "#/components/schemas/History"
                },
                "dx": {
                  "nullable": true,
                  "type": "object",
                  "description": "前端已生成诊断对象（复用）；缺省服务端重算；红旗一律以服务端规则重算为准"
                }
              }
            }
          }
        }
      }
    },
    "responses": {
      "JsonEnvelope": {
        "description": "统一壳 {code:0,data} + x-request-id",
        "headers": {
          "x-request-id": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Envelope"
            }
          }
        }
      },
      "NotFound": {
        "description": "未知病例 id——医生可理解文案",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "入站边界拒绝：请求体/问诊条数/单条字数超上限（阈值单一源 = frontend/tests/fixtures/request_limits.json）；文案不含内部阈值",
        "headers": {
          "x-request-id": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 413
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "BadJson": {
        "description": "请求体不是合法 JSON 对象——不再静默按空对象处理（v1.17.0 起显式 400）",
        "headers": {
          "x-request-id": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "BadShape": {
        "description": "JSON 合法但结构不合契约（如 history 非数组、元素非对象、dx 非对象）——双端同码同文案，v1.19.0 起权威面不再让其穿透到引擎报 500（消台账#28）；文案含故障编号供支持侧对账",
        "headers": {
          "x-request-id": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 422
                },
                "message": {
                  "type": "string",
                  "example": "请求参数不完整，请刷新后重试（故障编号 090fa889）"
                }
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "服务暂不可用——仅故障编号与重试文案，不含堆栈/内部路径",
        "headers": {
          "x-request-id": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Envelope"
            }
          }
        }
      }
    },
    "schemas": {
      "ChatResult": {
        "type": "object",
        "properties": {
          "conversation_id": {
            "type": "string",
            "nullable": true
          },
          "intent": {
            "type": "string",
            "enum": [
              "red_flag",
              "general_medical",
              "refund",
              "order_query",
              "tech_support",
              "out_of_scope"
            ]
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 0.95
          },
          "answer": {
            "type": "object",
            "properties": {
              "text": {
                "type": "string"
              },
              "citations": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "evidence_id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "source": {
                      "type": "string"
                    },
                    "year": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "red_flag": {
            "type": "object",
            "nullable": true,
            "properties": {
              "name": {
                "type": "string"
              },
              "severity": {
                "type": "string"
              },
              "advice": {
                "type": "string"
              }
            }
          },
          "handoff": {
            "type": "object",
            "nullable": true,
            "properties": {
              "reason_code": {
                "type": "string"
              },
              "reason_text": {
                "type": "string"
              },
              "ticket_id": {
                "type": "string"
              }
            }
          },
          "abstain": {
            "type": "boolean"
          },
          "mode": {
            "type": "string",
            "enum": [
              "deterministic",
              "live",
              "mock-fallback"
            ]
          },
          "answer_source": {
            "type": "string",
            "enum": [
              "deterministic",
              "agent"
            ]
          },
          "persisted": {
            "type": "boolean"
          },
          "persist_reason": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
      "ConversationDetail": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean"
          },
          "found": {
            "type": "boolean"
          },
          "conversation": {
            "type": "object",
            "nullable": true
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "intent_events": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "handoffs": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "feedback": {
            "type": "object",
            "nullable": true
          }
        }
      },
      "Envelope": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "description": "0=成功；失败时等于 HTTP 状态"
          },
          "data": {},
          "message": {
            "type": "string",
            "description": "医生可理解文案"
          }
        }
      },
      "History": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "role": {
              "type": "string",
              "enum": [
                "user",
                "assistant"
              ]
            },
            "content": {
              "type": "string"
            }
          }
        }
      },
      "Diagnosis": {
        "type": "object",
        "properties": {
          "primary": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "prob": {
                  "type": "string",
                  "enum": [
                    "高优先级",
                    "需鉴别",
                    "低可能"
                  ]
                },
                "strength": {
                  "type": "string",
                  "enum": [
                    "high",
                    "mid",
                    "low"
                  ]
                },
                "reasons": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "evidence_ids": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "仅限知识库白名单 kb-001~055"
                },
                "refs": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "differential": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "evidence": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "source": {
                  "type": "string"
                },
                "year": {
                  "type": "string"
                },
                "url": {
                  "type": "string"
                },
                "icd": {
                  "nullable": true,
                  "type": "string",
                  "description": "WHO ICD-10 初筛映射，未经临床复核不得作编码依据"
                },
                "score": {
                  "type": "number"
                }
              }
            }
          },
          "flags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "flag_details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "severity": {
                  "type": "string"
                },
                "advice": {
                  "type": "string"
                }
              }
            }
          },
          "mode": {
            "type": "string",
            "enum": [
              "live",
              "rule-fallback"
            ]
          },
          "fallback_reason": {
            "type": "string"
          },
          "fallback_cause": {
            "type": "string",
            "description": "降级原因分类（live 时为空串）：no_key/timeout/net_error/empty/bad_json/schema/truncated/content_filter/finish_unrecognized/unknown 或 http_<status>。只到类别，不含堆栈与内部路径。单一源 tests/fixtures/llm_fallback_causes.json"
          },
          "faq": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "trace": {
            "type": "object",
            "properties": {
              "evidence_ids": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "rounds": {
                "type": "integer"
              },
              "symptoms": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "state": {
            "type": "object"
          },
          "fhir": {
            "type": "object",
            "description": "FHIR R4（light 子集）Bundle：Patient/Encounter/Condition/Observation/DiagnosticReport，供 HIS 集成。只读派生视图，不参与决策链；零时钟字段保证双端逐字节对账；术语绑定仅用 HL7 已发布 CodeSystem，本地语义走本仓命名空间。判据见 tests/fhir_guard.mjs。"
          },
          "abstain": {
            "type": "boolean",
            "description": "第三态开关：true 表示本次证据不足或输入超出范围，未给出倾向性诊断（v1.25.0 起，台账#52）"
          },
          "scope_status": {
            "type": "string",
            "enum": [
              "in-scope",
              "insufficient-information",
              "out-of-scope"
            ],
            "description": "适用范围状态。红旗命中时一律 in-scope——弃权永远不得吞掉危险信号"
          },
          "top_evidence_score": {
            "type": "number",
            "description": "本次最高 BM25 证据分，与 ABSTAIN_T（双端同值，由 tests/ood_probe.mjs 实算对账）比较得出 abstain"
          },
          "abstain_reason": {
            "type": "string",
            "description": "给医生看的弃权理由，必须含「医生」主导口径；红旗与 evidence 在弃权时照常返回"
          },
          "scope_rule": {
            "type": [
              "string",
              "null"
            ],
            "description": "命中的能力级适用范围规则 id（权威在仓内 data/scope_rules.json，逐条带 rationale）；未命中为 null",
            "enum": [
              "imaging_or_report_reading",
              "weight_based_dosing",
              "non_human_patient",
              null
            ]
          }
        }
      }
    }
  }
}